feat(ui): moves confirm step to steps dir
