+++ app/src/components/custom-radio-buttons.tsx
@@ -1,0 +1,49 @@
+import { Box, Icon, useRadio, UseRadioProps } from "@chakra-ui/react";
+import { InventoryButtonCheckIcon } from "./icons";
+
+interface CustomRadioProps extends UseRadioProps {
+  children: React.ReactNode;
+}
+
+export function CustomRadio(props: CustomRadioProps) {
+  const { getInputProps, getCheckboxProps } = useRadio(props);
+
+  return (
+    <Box as="label">
+      <input {...getInputProps()} hidden />
+      <Box
+        {...getCheckboxProps()}
+        cursor={props.isDisabled ? "not-allowed" : "pointer"}
+        w="181px"
+        h="56px"
+        borderRadius="full"
+        display="flex"
+        justifyContent="center"
+        alignItems="center"
+        fontFamily="heading"
+        fontStyle="500"
+        textTransform="uppercase"
+        lineHeight="20px"
+        gap="8px"
+        letterSpacing="wide"
+        className="transition-all duration-150"
+        borderWidth={props.isChecked ? "0" : "1px"}
+        borderColor={props.isChecked ? "green.500" : "border.neutral"}
+        bg={props.isChecked ? "background.neutral" : "base.light"}
+        color={props.isChecked ? "content.link" : "content.secondary"}
+        _checked={{
+          bg: "background.neutral",
+          color: "content.link",
+          borderWidth: "1px",
+          borderColor: "interactive.secondary",
+        }}
+        _focus={{
+          boxShadow: "outline",
+        }}
+      >
+        {props.isChecked ? <Icon as={InventoryButtonCheckIcon} /> : ""}
+        {props.children}
+      </Box>
+    </Box>
+  );
+}
